Se finaliza el cambio hacia una screaming architecture
