+++ src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import Layout from './core/Layout.tsx'
-import Login  from './components/auth/Login.tsx'
+import Login from './auth/Login.tsx'
 import { createBrowserRouter, RouterProvider, Navigate} from 'react-router-dom'
-import BranchesPage from './pages/BranchesPage.tsx'
+import BranchesPage from './sucursales/pages/BranchesPage.tsx'
 
 const router = createBrowserRouter([
   {
